+++ network/network.go
@@ -91,8 +91,8 @@ type Network interface {
 	// MaxBatch returns the maximum batch size for network responses
 	MaxBatch() uint64
 
-	// BroadcastMainChannel broadcasts the given msg on main channel
-	BroadcastMainChannel(msg *proto.SignedMessage) error
+	// BroadcastMainTopic broadcasts the given msg on main channel
+	BroadcastMainTopic(msg *proto.SignedMessage) error
 
 	// SubscribeToMainTopic subscribes to main topic
 	SubscribeToMainTopic() error
